+++ Cesium3DTilesSelection/include/Cesium3DTilesSelection/CreditSystem.h
@@ -44,12 +44,22 @@ public:
    * @return If this string already exists, returns a Credit handle to the
    * existing entry. Otherwise returns a Credit handle to a new entry.
    */
-  Credit createCredit(const std::string& html);
+  Credit createCredit(const std::string& html, bool showOnScreen = false);
+
+  /**
+   * @brief Gets whether or not the credit should be shown on screen.
+   */
+  bool shouldBeShownOnScreen(Credit credit) const noexcept;
 
   /**
    * @brief Get the HTML string for this credit
    */
   const std::string& getHtml(Credit credit) const noexcept;
+
+  /**
+   * @brief Get whether this credit is the Cesium logo.
+   */
+  bool isIon(Credit credit) const noexcept;
 
   /**
    * @brief Adds the Credit to the set of credits to show this frame
@@ -82,7 +92,9 @@ private:
 
   struct HtmlAndLastFrameNumber {
     std::string html;
+    bool showOnScreen;
     int32_t lastFrameNumber;
+    bool isIon;
     int count = 0;
   };