+++ Cesium3DTilesSelection/src/Tile.cpp
@@ -543,7 +543,8 @@ void Tile::processLoadedContent() {
       // else if it has a model, try to get the copyright
       else if (this->_pContent->model->asset.copyright) {
         this->_pContent->credit = externals.pCreditSystem->createCredit(
-            *this->_pContent->model->asset.copyright);
+            *this->_pContent->model->asset.copyright,
+            this->getTileset()->getOptions().showCreditsOnScreen);
       }
 
       // A new and improved bounding volume.