+++ CHANGES.md
@@ -12,6 +12,7 @@
 - Add ScopeGuard utility to automatically execute function when exiting a scope.
 - Add glTF copyright information to the list of tileset credits.
 - Credits are now sorted based on the number of occurrences.
+- Added option to show credits on screen.
 
 ### v0.13.0 - 2022-03-01